TST: split up pandas/tests/test_multilevel.py (#23797)
